Resolve the bucket's external name in the function
This essentially does what a bucketRef would do, but inside the function
code. This makes the function more complex, but I consider it a best
practice.

Per the linked Crossplane issue, I don't think it's tenable to continue
down the path of implementing every possible cross-resource reference at
the provider level. It's much more flexible to just do it in a function.

https://github.com/crossplane/crossplane/issues/4627

Signed-off-by: Nic Cope <nicc@rk0n.org>