Update KCL function and tests to be more idiomatic
Stop setting the `metadata.name` of resources in the KCL function so that we let
Crossplane generate a name that will avoid conflicts. To accommodate this in the
tests, split the composition test in two. The first test is for the case where
the bucket has not yet been named; the second for the test where it has. We now
match desired resources by the `composition-resource-name` annotation since we
don't name them.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>